script.cpp fix

+++ src/script/script.cpp
@@ -576,7 +576,7 @@ bool EvalScript(vector<vector<unsigned char> >& stack, const CScript& script, co
                 case OP_DEPTH:
                 {
                     // -- stacksize
-                    CBigNum bn(stack.size());
+                    CBigNum bn((int)stack.size());
                     stack.push_back(bn.getvch());
                 }
                 break;
@@ -732,7 +732,7 @@ bool EvalScript(vector<vector<unsigned char> >& stack, const CScript& script, co
                     // (in -- in size)
                     if (stack.size() < 1)
                         return false;
-                    CBigNum bn(stacktop(-1).size());
+                    CBigNum bn((uint32_t)stacktop(-1).size());
                     stack.push_back(bn.getvch());
                 }
                 break;
